erasure_code: Add utilities to merge extent lists in ECCommon
More preparation for future work. Future work will require that multiple
reads to extents of shards will need to be combined into larger reads.

In particular, consider the case where one shard has failed. The reads
for that shard need to be copied to every other shard and the parity
shards in order to recover the data. This commit provides utilities to
minimise the number of required reads.