erasure-code/test: Allow TestEcBackend.cc to call non-static functions.
Previously all tests in this file were constrained by the need to
call statics, since insufficient stubs existed to construct an
ECReadPipeLine object.

This commit adds such a structure and adds a trivial (and duplicated) test to prove that a non-static function cal be called.

This function is part of ground work being put in place for some
more extensive changes being made to the EC code to make it more
efficient for block and file workloads.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>