osd: Update PGBackend interface to meet requirements of new optimised code.
In the new EC optimisations, we are dropping the hinfo attribute for some plugins/pool settings.  As a result, the EC code cannot calculate the object size on its own. The object_read_async therefore passes in the object size (already known outside EC) so that EC can perform the necessary RMW calculations without hinfo.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>