erasure-code: Rewrite the encode_chunks interface for all plugins
We have changed the encode_chunks interface to take an in map and an out map.
The in map contains the data shards to be encoded. The out map contains the
empty buffers that the plugin will write the parity data to.
The old non-optimized EC path has been updated to call the new encode_chunks
interface from erasurecode.cc
The new EC optimizations path calls the encode_chunks interface from ECUtils.

Signed-off-by: Jamie Pryde <jamiepry@uk.ibm.com>